backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
Pillow==10.1.0
django-filter==23.3
markdown==3.5.1
dj-rest-auth==5.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.5.0
